+++ actor-example/actors/src/fake_temperature_sensor.rs
@@ -1,0 +1,98 @@
+use rtic_actor_traits::{Post, Receive};
+
+use crate::{DoTemperatureRead, TemperatureReadingCelsius};
+
+pub struct FakeTemperatureSensor<P>
+where
+    P: Post<TemperatureReadingCelsius>,
+{
+    delta: i32,
+    outbox: P,
+    temperature: i32,
+}
+
+// a real temperature sensor would use the embedded-hal traits (e.g. I2C) or some higher level trait
+impl<P> FakeTemperatureSensor<P>
+where
+    P: Post<TemperatureReadingCelsius>,
+{
+    pub fn new(outbox: P, initial_temperature: i32, delta: i32) -> Self {
+        Self {
+            delta,
+            outbox,
+            temperature: initial_temperature,
+        }
+    }
+}
+
+impl<P> Receive<DoTemperatureRead> for FakeTemperatureSensor<P>
+where
+    P: Post<TemperatureReadingCelsius>,
+{
+    fn receive(&mut self, _: DoTemperatureRead) {
+        self.outbox
+            .post(TemperatureReadingCelsius(self.temperature))
+            .expect("OOM");
+        self.temperature += self.delta;
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use rtic_post_spy::PostSpy;
+
+    use super::*;
+
+    #[test]
+    fn on_read_it_posts_reading() {
+        let mut sensor = FakeTemperatureSensor::new(PostSpy::default(), 0, 0);
+
+        // manually send a message
+        let message = DoTemperatureRead;
+        sensor.receive(message);
+
+        let spy = sensor.outbox;
+        let posted_messages = spy.posted_messages::<TemperatureReadingCelsius>();
+        assert_eq!(1, posted_messages.count());
+    }
+
+    #[test]
+    fn reading_starts_at_initial_temperature() {
+        let initial_temperature = 1;
+        let mut sensor = FakeTemperatureSensor::new(PostSpy::default(), initial_temperature, 0);
+
+        // manually send a message
+        let message = DoTemperatureRead;
+        sensor.receive(message);
+
+        let spy = sensor.outbox;
+        let mut posted_messages = spy.posted_messages::<TemperatureReadingCelsius>();
+        assert_eq!(
+            Some(&TemperatureReadingCelsius(initial_temperature)),
+            posted_messages.next()
+        );
+    }
+
+    #[test]
+    fn reading_changes_by_delta() {
+        let initial_temperature = 42;
+        let delta = 1;
+        let mut sensor = FakeTemperatureSensor::new(PostSpy::default(), initial_temperature, delta);
+
+        // manually send a message
+        let message = DoTemperatureRead;
+        sensor.receive(message);
+        sensor.receive(message);
+
+        let spy = sensor.outbox;
+        let mut posted_messages = spy.posted_messages::<TemperatureReadingCelsius>();
+        assert_eq!(
+            Some(&TemperatureReadingCelsius(initial_temperature)),
+            posted_messages.next()
+        );
+        assert_eq!(
+            Some(&TemperatureReadingCelsius(initial_temperature + delta)),
+            posted_messages.next()
+        );
+    }
+}